+++ web/client/src/library/components/editor/hooks.ts
@@ -29,7 +29,6 @@ import {
   SQLMeshDialect,
   SQLMeshDialectCleanUp,
 } from './extensions/SQLMeshDialect'
-import { Table } from 'apache-arrow'
 
 export {
   useDefaultExtensions,
@@ -231,31 +230,4 @@ function useSQLMeshModelExtensions(
 
 function useSQLMeshDialect(): [ExtensionSQLMeshDialect, Callback] {
   return [SQLMeshDialect, SQLMeshDialectCleanUp]
-}
-
-export function toTableRow(
-  row: Array<[string, TableCellValue]> = [],
-): Record<string, TableCellValue> {
-  // using Array.from to convert the Proxies to real objects
-  return Array.from(row).reduce(
-    (acc, [key, value]) => Object.assign(acc, { [key]: value }),
-    {},
-  )
-}
-
-type TableCellValue = number | string | null
-type TableRows = Array<Record<string, TableCellValue>>
-type TableColumns = string[]
-type ResponseTableColumns = Array<Array<[string, TableCellValue]>>
-
-export function getTableDataFromArrowStreamResult(
-  result: Table<any>,
-): [TableColumns?, TableRows?] {
-  if (result == null) return []
-
-  const data: ResponseTableColumns = result.toArray() // result.toArray() returns an array of Proxies
-  const rows = Array.from(data).map(toTableRow) // using Array.from to convert the Proxies to real objects
-  const columns = result.schema.fields.map(field => field.name)
-
-  return [columns, rows]
 }
